Bodrov Daniil. Lab 4. Option 4. (#160)
I have implemented a pass in MLIR that counts the number of times each function is called by other functions within the module. The goal was to attach this count as an attribute to each function operation. To achieve this, we iterated through all operations in the module, tracking function calls and their frequencies. Then, we attached the call count as an attribute to each function operation. Finally, we registered this pass to be accessible as a plugin.