Kulagin Aleksandr / lab 4 / var 2 (#159)
Has features:

A support for d = a - b * c and d = b * c - a.
If the result of a fadd or fsub operation is not used anywhere else, we can safely delete it.
Issues:

Works correctly with -O1+. With -O0 I cannot bypass store operation because I can't catch it with getDefiningOp. I can catch load operation, but it's operand is not a result of fadd. It's a result of alloca, so it's useless. I hope it's okay.